Reconfigures movement
Changes movement to be restricted by the canMove and canJump functions
and not within handleClick.
The canMove and canJump functions now return an array of possible moves
or null if none are available.
This allows slimmer code within handleClick and should allow future
highlighting of potential moves and checking a "no more moves available"
win condition.