Fixes naming error and alters comments
Minor update which changes the values of upRight, upLeft, jumpUpRight,
and jumpUpLeft in the canMove and canJump methods to reflect their
logical equivalent. I.e. upRight now no longer represents index - 9,
which should point to the upper left square on the board.

Adds a few comments as a reminder to remove test code, removes comments
explaining the above logical incongruency, and alters comments for the
test board section to allow for collapsing of entire section.