Adds highlighting for possible moves.
Adds a function to check for all possible moves for a piece and return
an array with the first element being the index of the selected piece
and all other elements being indices of those possible moves.

Adds a class name to those possible moves on render. This combined with
a basic style added to index.css allows for highlighting possible moves.